Struct abyss::dev::pci::virtio::block::VirtIoBlockCfg

pub struct VirtIoBlockCfg(_);

Device configuration layout

The capacity of the device (expressed in 512-byte sectors) is always present. The availability of the others all depend on various feature bits as indicated above. The parameters in the configuration space of the device max_discard_sectors discard_sector_alignment are expressed in 512-byte units if the VIRTIO_BLK_F_DISCARD feature bit is negotiated. The max_write_zeroes_sectors is expressed in 512-byte units if the VIRTIO_BLK_F_WRITE_ZEROES feature bit is negotiated.

Implementations

impl VirtIoBlockCfg

pub fn new_from_mmio_area(area: MmioArea) -> Self

Create new mmio area.

pub fn va(&self) -> Va

Get starting virtual address.

impl VirtIoBlockCfg

pub fn capacity(&self) -> MmioAccessor<u64, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn size_max(&self) -> MmioAccessor<u32, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn seg_max(&self) -> MmioAccessor<u32, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn geometry_cylinders(&self) -> MmioAccessor<u16, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn geometry_heads(&self) -> MmioAccessor<u8, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn geometry_sectors(&self) -> MmioAccessor<u8, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn blk_size(&self) -> MmioAccessor<u16, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn topology_physical_block_exp(&self) -> MmioAccessor<u8, { _ }, { _ }>

of blocks per physical block (log2)

impl VirtIoBlockCfg

pub fn topology_alignment_offset(&self) -> MmioAccessor<u8, { _ }, { _ }>

offset of first aligned logical block

impl VirtIoBlockCfg

pub fn topology_min_io_size(&self) -> MmioAccessor<u16, { _ }, { _ }>

suggested minimum I/O size in blocks

impl VirtIoBlockCfg

pub fn topology_opt_io_size(&self) -> MmioAccessor<u32, { _ }, { _ }>

optimal (suggested maximum) I/O size in blocks

impl VirtIoBlockCfg

pub fn writeback(&self) -> MmioAccessor<u8, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn max_discard_sectors(&self) -> MmioAccessor<u32, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn max_discard_seg(&self) -> MmioAccessor<u32, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn discard_sector_alignment(&self) -> MmioAccessor<u32, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn max_write_zeros_sectors(&self) -> MmioAccessor<u32, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn max_write_zeros_seg(&self) -> MmioAccessor<u32, { _ }, { _ }>

impl VirtIoBlockCfg

pub fn write_zeros_may_unmap(&self) -> MmioAccessor<u8, { _ }, { _ }>