Enum abyss::dev::pci::virtio::DeviceType

#[repr(u32)]
pub enum DeviceType {

    Reserved,
    NetworkCard,
    Block,
    Console,
    EntropySource,
    MemoryBallonningTranditional,
    IoMemory,
    RpMsg,
    ScsiHost,
    Transport9P,
    Mac802WLan,
    RprocSerial,
    VirtioCaif,
    MemoryBalloon,
    Gpu,
    TimerClock,
    Input,
    Socket,
    Crypto,
    SignalDistributionModule,
    Pstore,
    IoMmu,
    Memory,

}

Device Types.

On top of the queues, config space and feature negotiation facilities built into virtio, several devices are defined. The following device IDs are used to identify different types of virtio devices. Some device IDs are reserved for devices which are not currently defined in this standard. Discovering what devices are available and their type is bus-dependent.

Variants

Reserved

NetworkCard

Block

Console

EntropySource

MemoryBallonningTranditional

IoMemory

RpMsg

ScsiHost

Transport9P

Mac802WLan

RprocSerial

VirtioCaif

MemoryBalloon

Gpu

TimerClock

Input

Socket

Crypto

SignalDistributionModule

Pstore

IoMmu

Memory