Struct abyss::dev::pci::virtio::pci::VirtIoPciCommonCfg

pub struct VirtIoPciCommonCfg(_);

4.1.4.3 Common configuration structure layout

Implementations

impl VirtIoPciCommonCfg

pub fn new_from_mmio_area(area: MmioArea) -> Self

Create new mmio area.

pub fn va(&self) -> Va

Get starting virtual address.

impl VirtIoPciCommonCfg

pub fn device_features_select(&self) -> MmioAccessor<u32, { _ }, { _ }>

The driver uses this to select which feature bits device_feature shows. Value 0x0 selects Feature Bits 0 to 31, 0x1 selects Feature Bits 32 to 63, etc.

impl VirtIoPciCommonCfg

pub fn device_features(&self) -> MmioAccessor<u32, { _ }, { _ }>

The device uses this to report which feature bits it is offering to the driver: the driver writes to device_feature_select to select which feature bits are presented.

impl VirtIoPciCommonCfg

pub fn driver_features_select(&self) -> MmioAccessor<u32, { _ }, { _ }>

The driver uses this to select which feature bits driver_feature shows. Value 0x0 selects Feature Bits 0 to 31, 0x1 selects Feature Bits 32 to 63, etc.

impl VirtIoPciCommonCfg

pub fn driver_features(&self) -> MmioAccessor<u32, { _ }, { _ }>

The driver writes this to accept feature bits offered by the device. Driver Feature Bits selected by driver_feature_select.

impl VirtIoPciCommonCfg

pub fn msix_config(&self) -> MmioAccessor<u16, { _ }, { _ }>

The driver sets the Configuration Vector for MSI-X.

impl VirtIoPciCommonCfg

pub fn num_queues(&self) -> MmioAccessor<u16, { _ }, { _ }>

The device specifies the maximum number of virtqueues supported here.

impl VirtIoPciCommonCfg

pub fn device_status(&self) -> MmioAccessor<Status, { _ }, { _ }>

The driver writes the device status here (see 2.1). Writing 0 into this field resets the device.

impl VirtIoPciCommonCfg

pub fn config_generation(&self) -> MmioAccessor<u8, { _ }, { _ }>

Configuration atomicity value. The device changes this every time the configuration noticeably changes.

impl VirtIoPciCommonCfg

pub fn queue_select(&self) -> MmioAccessor<u16, { _ }, { _ }>

Queue Select. The driver selects which virtqueue the following fields refer to.

impl VirtIoPciCommonCfg

pub fn queue_size(&self) -> MmioAccessor<u16, { _ }, { _ }>

Queue Size. On reset, specifies the maximum queue size supported by the device. This can be modified by the driver to reduce memory requirements. A 0 means the queue is unavailable.

impl VirtIoPciCommonCfg

pub fn queue_msix_vector(&self) -> MmioAccessor<u16, { _ }, { _ }>

The driver uses this to specify the queue vector for MSI-X.

impl VirtIoPciCommonCfg

pub fn queue_enable(&self) -> MmioAccessor<u16, { _ }, { _ }>

The driver uses this to selectively prevent the device from executing requests from this virtqueue. 1 - enabled; 0 - disabled.

impl VirtIoPciCommonCfg

pub fn queue_notify_off(&self) -> MmioAccessor<u16, { _ }, { _ }>

The driver reads this to calculate the offset from start of Notification structure at which this virtqueue is located. Note: this is not an offset in bytes. See 4.1.4.4 below.

impl VirtIoPciCommonCfg

pub fn queue_desc(&self) -> MmioAccessor<u64, { _ }, { _ }>

The driver writes the physical address of Descriptor Area here. See section 2.5.

impl VirtIoPciCommonCfg

pub fn queue_driver(&self) -> MmioAccessor<u64, { _ }, { _ }>

The driver writes the physical address of Driver Area here. See section 2.5.

impl VirtIoPciCommonCfg

pub fn queue_device(&self) -> MmioAccessor<u64, { _ }, { _ }>

The driver writes the physical address of Device Area here. See section 2.5.

impl VirtIoPciCommonCfg

pub fn get_device_features(&self) -> u64

pub fn set_features(&self, features: u64)