abyss

Struct Pa

#[repr(transparent)]
pub struct Pa(usize);

Represents a physical address.

The Pa (Physical Address) struct is a wrapper around the usize type, which represents a physical address in memory. It is used to handle addresses that correspond directly to the hardware memory locations in the physical address space.

This struct provides methods to:

Create a new physical address with validation.

Convert a physical address to a virtual address.

Convert a physical address to a raw usize type.

Example:

let pa = Pa::new(0x1234_5678_9ABC_DEF0).unwrap();
let kva = pa.into_kva(); // Convert to a kernel virtual address.

Tuple Fields
0: usize
Implementations

impl Pa

pub const ZERO: Self

The physical address 0.

This constant represents the special address 0.

pub const fn new(addr: usize) -> Option<Self>

Creates a new physical address if the address is valid.

This method attempts to create a new Pa instance by validating the provided physical address. The address must be less than 0xffff_0000_0000_0000, which ensures it falls within the valid physical address range.

Arguments

addr: A usize representing the physical address.

Returns

Some(Pa) if the address is valid.

None if the address is outside the valid range.

Example:

let pa = Pa::new(0x1234_5678_9ABC_DEF0);

pub const fn into_usize(self) -> usize

Cast the physical address into a raw usize.

This method allows the physical address to be cast into a raw usize value, which can be used for low-level operations like pointer arithmetic or addressing.

Returns

The underlying usize value representing the physical address.

pub const fn into_kva(self) -> Kva

Convert the physical address to a virtual address.

This method allows you to convert a Pa (physical address) to a [Kva] (kernel virtual address). The conversion uses a fixed offset to transform the physical address into a corresponding kernel virtual address.

Returns

The corresponding kernel virtual address as a [Kva] instance.

Example:

let pa = Pa::new(0x1234_5678_9ABC_DEF0).unwrap();
let va = pa.into_kva();  // Convert the physical address to kernel virtual address

pub const fn page_down(self) -> Self

Align down the physical address to the page boundary.

pub const fn page_up(self) -> Self

Align up to the physical address to the page boundary.

pub const fn offset(self) -> usize

Extracts the page offset from the physical address.

This method retrieves the lower bits of the address that represent the offset within a memory page. The offset is useful when working with memory operations that need to determine the position within a page.

Returns

The offset within the page as a usize.

Example

let pa = Pa::new(0x1234_5678).unwrap();
let offset = pa.offset();
assert_eq!(offset, 0x678); // Example offset within the page